gleason-thesis-repo/requirements.txt
numpy
pandas
matplotlib
scipy
librosa
soundfile
openai-whisper
resemblyzer
scikit-learn
statsmodels
torch
transformers
datasets
praat-parselmouth
pydub
sqlalchemy
